Added authorization to business and category routes

+++ src/routes/categories.js
@@ -2,14 +2,16 @@
 
 const express = require('express');
 const router = express.Router();
+const jwt = require('jsonwebtoken');
 
+const auth = require('../auth/auth-middleware');
 const Model = require('../models/category/category');
 const modelRepository = new Model();
 
 router.get('/categories', wrap(handleGet));
-router.post('/category', wrap(handlePost));
-router.put('/category/:id', wrap(handlePut));
-router.delete('/category/:id', wrap(handleDelete));
+router.post('/category', auth, wrap(handlePost));
+router.put('/category/:id', auth, wrap(handlePut));
+router.delete('/category/:id', auth, wrap(handleDelete));
 
 /**
  * Wraps a route callback with a try/catch, which passes on uncaught errors to be properly handled
@@ -79,6 +81,7 @@ async function handleGet(req, res, next) {
  * @returns {error} 500 - Internal Server Error
  */
 async function handlePost(req, res, next) {
+  verifyAdmin(req);
   let record = await modelRepository.post(req.body);
   send(record, res, 201);
 }
@@ -94,6 +97,7 @@ async function handlePost(req, res, next) {
  * @returns {error} 500 - Internal Server Error
  */
 async function handlePut(req, res, next) {
+  verifyAdmin(req);
   const record = await modelRepository.put(req.params.id, req.body);
   verifyExists(record, req.params.id);
   send(record, res);
@@ -109,9 +113,22 @@ async function handlePut(req, res, next) {
  * @returns {error} 500 - Internal Server Error
  */
 async function handleDelete(req, res, next) {
+  verifyAdmin(req);
   let record = await modelRepository.delete(req.params.id);
   verifyExists(record, req.params.id);
   send(record, res);
 }
 
+function get401() {
+  const error = new Error('You are not authorized to do this action');
+  error.status = 401;
+  return error;
+}
+
+function verifyAdmin(req) {
+  const tokenData = jwt.decode(req.token);
+  if (tokenData.type !== 'admin') {
+    throw get401();
+  }
+}
 module.exports = router;
